Force conffiles to be overwritten
This fixes a bug where the default conffile handling caused dpkg to
propmpt for input and abort the upgrade. Read about the default behavior
in dpkg(1).

This also simplifies the sometimes subtle conffile rules and makes the
results of this upgrade mechanism less surprising. Now, when a package
gets installed, so does the packaged configuration, regardless of
whether the configuration had been locally modified or deleted.